Make enterprise licensed
